Updated Ceilometer HA
When Ceilometer HA was setup, the HAProxy config was setup with an
HTTP check. This is not possible in Ceilometer without the HTTP
request having authentication headers as ceilometer will return
a 401 and fail the HAProxy check.

The check type has been changed to TCP and the balancing type was
set to source with no options.